pbl/requirements.txt
captcha==0.3
dj-database-url==0.5.0
dj-static==0.0.6
Django~=2.2.4
gunicorn==20.0.4
html5lib==1.1
Pillow==6.1.0
PyMySQL==0.9.3
PyPDF2==1.26.0
pytz==2019.2
reportlab==3.5.42
six==1.15.0
sqlparse==0.3.0
static3==0.7.0
virtualenv==16.7.5
virtualenvwrapper-win==1.2.5
webencodings==0.5.1
xhtml2pdf==0.2.4
django-easy-pdf~=0.1.1